Populert event_types tabell

+++ src/main/resources/db/migration/V2__create_event_types_table.sql
@@ -1,5 +1,5 @@
 CREATE TABLE "event_types" (
-    "event_type_id" SERIAL PRIMARY KEY,
+    "event_type_id" NUMERIC(4,0) PRIMARY KEY,
     "description" VARCHAR(256) NOT NULL,
     "status" event_status NOT NULL
 );
